refactor(capabilities): merge http and http_server into one submodule owning their kinds
Collapse the sibling `http.rs` (egress client) and `http_server.rs`
(ingress server) into one `http/` directory submodule (`client.rs` +
`server.rs` + `kinds.rs` + `mod.rs`), and move the eight HTTP wire kinds
out of `aether-kinds` into the capability that owns them (ADR-0121). The
two stay distinct capabilities — separate cap structs, separate
`NAMESPACE` / mailboxes — merely co-located under one parent module.

`http/kinds.rs` is always-on and wasm-safe (`aether_data::{Kind, Schema}`
+ serde only), so `default-features = false` consumers keep compiling.
Kind names are unchanged, so the name+schema-derived wire ids are
unchanged. The roundtrip tests relocate alongside their kinds.

Repoints the intra-crate kind imports in `client.rs` / `server.rs`, the
flat re-exports in `aether-capabilities` `lib.rs`, the bundle callers
(`chassis_common.rs`, `cli.rs`), and the wasm fixture handler to the new
`http::kinds` path; drops the moved kinds from the `aether-kinds`
descriptor-coverage test.
