+++ crates/aether-kinds/src/descriptors.rs
@@ -52,8 +52,8 @@ mod tests {
 
     use crate::{
         CliSend, CliSendResult, Delete, DeleteResult, DrawTriangle, DropComponent, DropResult,
-        Fetch, FetchResult, FsFetch, FsFetchResult, Key, LifecycleUnsubscribeAll, List, ListResult,
-        LoadComponent, LoadResult, LyriaGenerate, LyriaGenerateResult, Mat4Apply, MessagesSend,
+        FsFetch, FsFetchResult, Key, LifecycleUnsubscribeAll, List, ListResult, LoadComponent,
+        LoadResult, LyriaGenerate, LyriaGenerateResult, Mat4Apply, MessagesSend,
         MessagesSendResult, MouseButton, MouseMove, NanobananaGenerate, NanobananaGenerateResult,
         NoteOff, NoteOn, Ping, Pong, ProcessExited, Read, ReadResult, RecordResult,
         ReplaceComponent, ReplaceResult, SetMasterGain, Spawn, SpawnResult, SubscribeInput,
@@ -124,8 +124,6 @@ mod tests {
         assert!(names.contains(&DeleteResult::NAME));
         assert!(names.contains(&List::NAME));
         assert!(names.contains(&ListResult::NAME));
-        assert!(names.contains(&Fetch::NAME));
-        assert!(names.contains(&FetchResult::NAME));
         assert!(names.contains(&FsFetch::NAME));
         assert!(names.contains(&FsFetchResult::NAME));
         assert!(names.contains(&Spawn::NAME));
